+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 New `functional` preset (included in `recommended`) with the following new rules:
 - [no-httpinterceptors-token](https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/functional/NO_HTTPINTERCEPTORS_TOKEN.md)
+- [no-withinterceptorsfromdi](https://github.com/cyrilletuzi/eslint-plugin-angular-modern/blob/main/docs/functional/NO_WITHINTERCEPTORSFROMDI.md)
 
 ## 1.4.0 (2026-08-17)